+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 
 ## Version 2
 
+### Version 2.2.1
+
+* [142](https://github.com/mlebreuil/netbox-contract/issues/142) Gives the option to enter contract yearly recuring costs instead of only monthly recuring costs.
+Corresponding value is used to calculate the invoices amount without rounding approximations.
+
 ### Version 2.2.0
 
 * [140](https://github.com/mlebreuil/netbox-contract/issues/140) Add the "Invoice line" and "Accounting dimension" models. In order to simplify invoices creation, it is possible to selsct one invoice as the template for each contract; Its accounting lines will automatically be copied to the new invoices for the contract. The amount of the first line will be updated so that the sum of the amount for each invoice line match the invoice amount.